WMI/BitLocker: if there is no Win32_EncryptableVolume, assume no BitLocker
Apparent regression from T21417 - previously we searched for the property and
drive at the same time, and took absence to mean no BitLocker. It seems there
are systems without BitLocker which have no Win32_EncryptableVolume object for
the system drive, which seems to suggest the whole feature is disabled or
missing somehow. Return FALSE in this case as before.

https://phabricator.endlessm.com/T22290