WMI: call GetConversionStatus
The ConversionStatus property is not [documented][0] to exist on
Win32_EncryptableVolume. In my testing, it does exist on Windows 10,
but [forum reports][1] indicate that it does not exist on Windows 7.
As a result, we always entered the fail-safe path of assuming the disk
is encrypted, and Endless OS could not be installed.

Instead, call [GetConversionStatus][2] and get ConversionStatus from the
object it returns. ProtectionStatus is [documented][0] to exist, and
the very helpful forum report agrees, so we continue to access that as
we were before.

GetConversionStatus actually is documented to accept one in-parameter,
uint32 PrecisionFactor, but that parameter itself is only documented as:

> A value from 0 to 4 that specifies the precision level

and calling the method without it works fine. (I guess it may affect the
accuracy of EncryptionPercentage?)

[0]: https://msdn.microsoft.com/en-us/library/windows/desktop/aa376483(v=vs.85).aspx
[1]: https://community.endlessos.com/t/problem-with-installation-drive-encrypted/6300/5
[2]: https://msdn.microsoft.com/en-us/library/windows/desktop/aa376433(v=vs.85).aspx

https://phabricator.endlessm.com/T22663