Uninstall: only call GetPhysicalFromDriveLetter() when needed
We don't need a handle to the physical device in the eosldr uninstall
code path.  This works around an issue where FSCTL_LOCK_VOLUME fails.
Since we don't call GetPhysicalFromDriveLetter() on the eosldr install
code path, this leads to a situation where users can install Endless OS,
but not uninstall it.

https://phabricator.endlessm.com/T22788