DownloadManager: update remote URLs after redirects
BG_HTTP_REDIRECT_POLICY_ALLOW_REPORT causes remote URLs to be updated
when following redirects. In principle, this would allow us to log the
remote URLs when they change, though that is not implemented here.

Having done this, we need to match existing transfers by the path
component of the URL; otherwise, after a redirect we would always
abort an existing in-progress download and start again from the
beginning since the URLs no longer match.

This implementation assumes that all URLs in the job share a prefix.
This wasn't true in IsCoding() mode but the index on
images.endlessm-sf.com it tries to hit doesn't exist, so it's hard to
get too excited about this.

We fix up all URLs in a matching existing job to point back to the
canonical server -- perhaps the previous transfer failed because the
particular mirror we hit was broken.

https://phabricator.endlessm.com/T23121