DownloadManager: log remote URLs at important points
There's no explicit notification of redirection; JobModification is
called when almost anything about the job changes, and
JobError/JobTransferred are called once on error/success.

We unequivocally log the URLs on success, and on various error paths.
To try to get them into the log nearer the start of the transfer, We
track job state changes and log them when moving to TRANSFERRING.

Empirically, this is good enough for a freshly-created job: the job
does not enter state TRANSFERRING until the Content-Length for all
files is known, which entails following any redirects. When resuming a
job, we reset the URLs, but BITS doesn't re-check Content-Length, so
redirects are followed file-by-file. Good enough though.

https://phabricator.endlessm.com/T23121