Don't lock physical volume handle
This parameter of GetPhysicalHandle(), bLockDrive, gets passed though to
GetHandle(), and controls whether the resulting handle to the physical
volume (ie the disk itself, rather than a partition within it) is locked
with FSCTL_LOCK_VOLUME.

GetPhysicalFromDriveLetter() is used when (un)installing the EFI
bootloader (to find an ESP and mount it) and when uninstalling GRUB from
the boot sector.

EFI (un)install
---------------

Here, we don't need exclusive access to the volume: we're just listing
its partitions to find an ESP. However, installations do fail at this
point in the wild:

    21:27:24 - EndlessUsbToolDlg.cpp:5529 CEndlessUsbToolDlg::SetupEndlessEFI
    21:27:24 - Operation OP_SETUP_DUALBOOT(13) with progress 98
    21:27:24 - EndlessUsbToolDlg.cpp:5558 CEndlessUsbToolDlg::GetPhysicalFromDriveLetter
    21:27:24 - Opened drive \\.\PHYSICALDRIVE0 for write access
    21:27:39 - Could not get exclusive access to device \\.\PHYSICALDRIVE0: [0x00000005] Access is denied.
    21:27:39 - EndlessUsbToolDlg.cpp:5572 Error on acquiring disk handle. (GLE=[5])
    21:27:39 - EndlessUsbToolDlg.cpp:5537 Error on acquiring disk handle. (GLE=[5])
    21:27:39 - EndlessUsbToolDlg.cpp:5191 Error on SetupEndlessEFI (GLE=[5])
    21:27:39 - SetupDualBoot exited with error.

https://phabricator.endlessm.com/T16194

MBR GRUB uninstall
------------------

In principle, we want exclusive access to the volume here: we're making
dangerous changes to its boot sector.  This has always worked reliably
for me, but it has been seen to fail in the wild:

    14:50:27 - c:\projects\rufus\src\endless\endlessusbtooldlg.cpp:6054 CEndlessUsbToolDlg::UninstallDualBoot
    14:50:27 - c:\projects\rufus\src\endless\endlessusbtooldlg.cpp:5566 CEndlessUsbToolDlg::IsLegacyBIOSBoot
    14:50:27 - c:\projects\rufus\src\endless\eosldr.cpp:61 EosldrInstaller::IsInstalled C:\
    14:50:27 - c:\projects\rufus\src\endless\endlessusbtooldlg.cpp:5679 CEndlessUsbToolDlg::GetPhysicalFromDriveLetter
    14:50:27 - Opened drive \\.\PHYSICALDRIVE0 for write access
    14:50:27 - Attempting FSCTL_LOCK_VOLUME for drive \\.\PHYSICALDRIVE0
    [… many duplicates elided …]
    14:50:27 - Attempting FSCTL_LOCK_VOLUME for drive \\.\PHYSICALDRIVE0
    14:50:27 - Could not get exclusive access to device \\.\PHYSICALDRIVE0: [0x00000005] Acesso negado.
    14:50:27 - c:\projects\rufus\src\endless\endlessusbtooldlg.cpp:5694 Error on acquiring disk handle. (GLE=[5])
    14:50:27 - c:\projects\rufus\src\endless\endlessusbtooldlg.cpp:6080 Error on acquiring disk handle. (GLE=[5])

I can't reproduce this on a Windows 10 1809 VM.  Perhaps it is dependent
on some anti-virus software?

It makes some sense that we cannot get exclusive access to the device,
because the mounted system partition is on it!  However, we really do
need to be able to write to it to put the original boot sector back.

Since June 2017, we have not created new dual-boot systems with GRUB as
the MBR (instead, we add GRUB to the Windows boot menu).  We don't
expect there are many such installations still in the wild.  Since the
uninstall entry in Windows' Add/Remove programs uses the version of this
tool that was used to install, this code path will only be reached when
a user is having trouble uninstalling Endless and has downloaded a new
version of the installer to try that.

This handle is only used to read and write the first 440 bytes of the
disk, and we do not expect anything else to be touching them in normal
operation of the system. So, grit our teeth and don't lock the handle.
This whole bootloader configuration is intrinsically scary and unsafe,
which is why we don't do it any more.

https://phabricator.endlessm.com/T26271