add 7zip-cpp
This we previously a git subtree, but that has proven to
be challanging for rebasing. Somehow files in the 7zip-cpp
directories were being treated as if they were in the project
root directory.

Instead, just copy in the contents from kjjones' repo.

It is hoped we can break this dependency eventually and remove
this from our tree entirely.