Allow failing DeletePartitions to be non-fatal
So, you can't DeletePartitons if there aren't any, or if they're not
recognized as such by windows. This means zeroed out devices or, much
more surprisingly, devices with our iso image flashed on them, can't be
re-flashed, as our process breaks at DeletePartitons.

Stop treating that error as fatal.

Honestly, I don't know why we need to DeletePartitions before we write
zeroes over the partition table anyway, but I'm cargo culting that in
from rufus, as the author has a great deal of experience dealing with the
weird and impossible world of windows volume management.

https://phabricator.endlessm.com/T29747