Zero the starts of partitions before setting up the partition table
In an attempt to stop windows from popping up windows explorer instances
when creating a USB stick, we now zero out the first 12KB of the partitions
we're about to create just before we create them.

12KB was chosen as it is large enough for exFAT's primary and secondary
boot blocks. It's much larger than required for FAT32/16/12, but isn't
large enough to take very long in practice.

Now there will never be a valid filesystem in any partition immediately
after we create the partition table, so windows has nothing to mount.

It will almost certainly still try to fire off the "please format this
partition" notifications, but those are already handled properly.

https://phabricator.endlessm.com/T29694