Delay move from combo usb creation to cancelled until ready
Upstream rufus moved to a more robust method of deleting partitions,
but it's dramatically slower. We use this function in combo usb
stick creation.

There's a long standing bug in our cancel path in which the UI moves
immediately to the cancelled screen, but CreateUSBStick thread
continues until it has a chance to check for cancellation.

Now that DeletePartitions() takes about 30 seconds to complete it's
entirely possible to hit the cancel button, then hit try again,
and get far enough into a retry that the cancel status is cleared
before CreateUSBStick gets past DeletePartitions and into a
cancel check.

Hilarity ensues.

This patch makes the GUI wait until CreateUSBStick checks for
cancellation and completes before updating to the canceled screen.

https://phabricator.endlessm.com/T29693