Perform ESP manipulations first
Windows 7 and 8 appear to hate USB sticks with multiple partitions, and
only see the "first" partition on the device.
(first is complicated here - but it appears that our switcheroo that
allows the third physical partition to be logically first still satisfies)

We need to move to a method where we only need to write to the first
partition at any time, and the partitioning code is a little simpler if
we do the ESP first.