Partition the drive twice so we only ever write to the first partition
Older windows versions apparently break down when they see USB devices
with multiple partitions. This results in things like the VDS partition
deletion code failing to complete, and windows internal tools like the
disk management snap-in being unable to mount some of the partitions on
a device. None of the mount code we depend on can successfully mount
these partitions either.

Work around this insanity with insanity of our own - partition the device
in two steps.

On the first step, partition with just the ESP, but call it a regular data
partition because it allows us to use "normal" interfaces to mount it. The
exact reason why this is mandatory and not just convenience will become
apparent in a later patch.

After performing all ESP duties, partition again, this time with the final
partition state in which the ExFAT partition is the first partition on the
device.

https://phabricator.endlessm.com/T29949