Use MountVolume instead of AltMountVolume for the soon-to-be-ESP
Since we've made the ESP temporarily be a regular data partition, we can
mount it using the MountVolume functions instead of AltMountVolume.

This is important because windows appears to mess with partition numbers,
and I've seen it get into a state where the ESP, even though it's the only
partition on the device, is reported as partition number 2 until the device
is removed and reinserted, at which point the number matches the value that
was written in the GPT.

That odd numbering confuses rufus' AltMountVolume code, which assumes that
the first partition in its array is partition 1, and manufactures a windows
device name (Harddisk1Partition1) for the wrong partition - image writing
fails when we are unable to mount that incorrect device name.

MountVolume sidesteps all of this nonsense by using the volume name which
we look up ourselves, and mounting that without having to come up with the
device name.

Window still has a wacky idea of what the partition is named, we just stop
caring.

Theoretically we could do the same thing for the ExFAT partition, but I've
reached an "it works, don't touch it!" point with this code, and will only
do that if it appears necessary at some point.