Update Endless company name in various places
On April 1st 2020, this application was transferred from Endless Mobile,
Inc. to Endless OS Foundation LLC.

I believe the strings in the .rc file are used in the Properties
dialog for the file in Windows Explorer. REGKEY_PUBLISHER_TEXT is used
when adding an entry to Change or Remove Programs to allow Endless OS to
be uninstalled from there.